+++ src/Handle.php
@@ -3,8 +3,9 @@ namespace Altmetric\Identifiers;
 
 class Handle
 {
-    public static function extract($str)
+    public static function extract(?string $str): array
     {
+        $str = $str ?? '';
         preg_match_all('#\b[\d.]+/\S+\b#u', $str, $matches);
 
         return $matches[0];